requirements.txt
Flask
PyJWT
Werkzeug
python-dotenv # For loading secret key from .env file